% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{add_residual}
\alias{add_residual}
\title{Add residual connection and project if feature dimensions do not match}
\usage{
add_residual(x, resid)
}
\description{
Add residual connection and project if feature dimensions do not match
}
