% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{apply_normalization}
\alias{apply_normalization}
\title{Applies specified normalization type to input x}
\usage{
apply_normalization(x, norm_type, depth = tail(shape_list2(x), 1), eps = 1e-09)
}
\description{
Applies specified normalization type to input x
}
