% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attention-utils.R
\name{combine_last_two_dimensions}
\alias{combine_last_two_dimensions}
\title{Reshape x so that the last two dimension become one.}
\usage{
combine_last_two_dimensions(x)
}
\description{
Reshape x so that the last two dimension become one.
}
