% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scratch.R
\name{compute_luong_score}
\alias{compute_luong_score}
\title{Vanilla without scale weight}
\usage{
compute_luong_score(
  query,
  keys,
  query_depth,
  scale = NULL,
  return_context = TRUE
)
}
\description{
Vanilla without scale weight
}
