% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attention-utils.R
\name{compute_qkv}
\alias{compute_qkv}
\title{query  [batch, length_q, channels]
memory [batch, length_m, channels] (optional, usually RNN hidden states)
return [batch, length, depth] (q, k ,v) tensors}
\usage{
compute_qkv(
  query,
  memory = NULL,
  key_depth = 64L,
  value_depth = 64L,
  q_filter_width = 1L,
  kv_filter_width = 1L,
  q_padding = "valid",
  kv_padding = "valid",
  vars_3d_num_heads = 0L
)
}
\description{
query  [batch, length_q, channels]
memory [batch, length_m, channels] (optional, usually RNN hidden states)
return [batch, length, depth] (q, k ,v) tensors
}
