% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{conv_relu_conv}
\alias{conv_relu_conv}
\title{Hidden layer with RELU activation followed by linear projection.}
\usage{
conv_relu_conv(
  x,
  filters,
  output_depth,
  first_kernel_size = 3L,
  second_kernel_size = 3L,
  padding = "SAME",
  dropout = 0
)
}
\description{
Hidden layer with RELU activation followed by linear projection.
}
