% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corral.R
\name{corral}
\alias{corral}
\title{Corral a character vector}
\usage{
corral(x, method = c("size", "asis", "name"), groups = NULL,
  collect = "Other")
}
\arguments{
\item{x}{A character vector (or any vector than can be coerced into a
character).}

\item{method}{A character string indicating the desired method of corralling
with \code{"size"} being the default.  This must be (an abbreviation of)
one of the strings \code{"size"}, \code{"asis"}, or \code{"name"}.  See
Details for more information.}

\item{groups}{Either NULL (all values are kept distinct),
a single number with the desired number of groups (floating numbers are truncated),
or a vector of values to keep distinct.  The default is NULL.}

\item{collect}{Either \code{NA} or a character string that denotes the name that the
"collected" values are given.  The default is "Other".}
}
\value{
The output of \code{corral} is a corralled factor vector with the
  same length as \code{x}.
}
\description{
\code{corral} returns a corralled factor vector.
}
\details{
\code{corral} returns a corralled vector \code{x}.
  \code{NA} values in \code{x} are not grouped during the corralling
  process but are preserved in the output.

\code{corral} is designed to be readable from the function call.
  For example:
\itemize{
  \item \code{corral(x, method="size", groups=5)} can be read as
    "\strong{corral} \strong{x} by \strong{size} into \strong{5} groups".
  \item \code{corral(x, method="asis", groups=c("a","b"))} can be read as
    "\strong{corral} \strong{x} \strong{as is} and keep only \strong{a} and
    \strong{b} distinct".
}

The output of \code{corral} is determined by the arguments \code{method} and
  \code{groups}.

\code{corral} offers a few different options for \code{method}:

\itemize{
  \item \strong{size}: The default option that corrals \code{x} based on
    the number of occurrences in \code{x}.
  \item \strong{asis}: Corrals \code{x} based on the order in which values are observed.
  \item \strong{name}: Corrals \code{x} based on alphanumerical order.
}

\code{corral} accepts either numeric or character values for \code{groups}:
\itemize{
  \item \strong{numeric}: Creates \code{groups} groups based on
    \code{method} and combines all other values into the \code{collect} category.
  \item \strong{character}: Creates a group for each value in \code{groups}
    and combines all other values into the \code{collect} category.
}

See the examples for some explicit illustration on how different combinations
  of \code{method} and \code{groups} result in different outputs.
}
\examples{
set.seed(1337)
x <- sample(letters, 1e4, replace=TRUE)
summary(x)

#####
# Common use cases
#

# I want to factorize by sample size!
x_all <- corral(x, "size")
summary(x_all)
# All values are kept and ordered by the number of occurrences

# I want to factorize by sample size but only have 5 values!
x_5 <- corral(x, "size", groups=5)
summary(x_5)
# The four most common values are kept and
# everything else is combined into "Other"

# I want to factorize but keep only specific values!
x_bar <- corral(x, "asis", groups=c("b", "a", "r"))
summary(x_bar)
# The values "b", "a", and "r" are explicitly kept and
# leveled based on the order provided (i.e. "b" then "a" then "r")

# I want to change the collected values to NA rather than Other!
x_NA <- corral(x, "asis", groups=c("b", "a", "r"), collect=NA)
summary(x_NA)
# The values "b", "a", and "r" are explicitly kept and
# leveled based on the order provided (i.e. "b" then "a" then "r")
}

