% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attention-utils.R
\name{embedding_to_padding}
\alias{embedding_to_padding}
\title{Calculates the padding mask based on which embeddings are all zero.}
\usage{
embedding_to_padding(emb)
}
\description{
emb Tensor with shape [..., depth]
}
\details{
Returns:
  a float Tensor with shape [...]. Each element is 1 if its 
  corresponding embedding vector is all zero, and is 0 otherwise.
}
