% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_transforms.R
\name{evaluate_transformations}
\alias{evaluate_transformations}
\title{Evaluate Transformations}
\usage{
evaluate_transformations(transformed_data)
}
\arguments{
\item{transformed_data}{A list of transformed datasets}
}
\value{
A data frame with skewness and kurtosis for each transformation
}
\description{
Evaluates different transformations based on normality measures (skewness and kurtosis).
}
