% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transform.R
\name{extracted}
\alias{extracted}
\title{Constructor function for \code{extracted} class}
\usage{
extracted(data)
}
\arguments{
\item{data}{A \code{data.frame} extracted from a database}
}
\value{
A \code{data.frame} with primary class \code{extracted}
}
\description{
Constructor function for \code{extracted} class
}

