% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateTransforms.R
\name{generateTransforms}
\alias{generateTransforms}
\title{Generating Transforms for Various Data Types}
\usage{
generateTransforms(formula, data, numeric = TRUE, character = TRUE,
  factor = TRUE, date = TRUE, ...)
}
\arguments{
\item{formula}{a \code{formula} to determine which variables to transform}

\item{data}{a \code{data.frame} to use for data transforms}

\item{numeric}{transform \code{numeric} data?}

\item{character}{transform \code{character} data?}

\item{factor}{transform \code{factor} data?}

\item{date}{transform \code{date} data?}

\item{...}{Other arguments passed to \code{\link[base]{summary}}}
}
\value{
Something
}
\description{
A suite of transforms for \code{numeric}, \code{character}, \code{factor}, and \code{date} data types.
}
\author{
Derek McCrae Norton <dnorton1@gmail.com>
}

