% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Impute missing values}
\usage{
impute(x, method, ...)
}
\arguments{
\item{x}{A vector.}

\item{method}{Either a value, a vector of values, or a function to use for imputation.
See Details for more information.}

\item{...}{Additional arguments for method when it is a function.}
}
\value{
The output of \code{impute} is the vector \code{x} with missing values imputed.
}
\description{
\code{impute} returns a vector with missing values imputed.
}
\details{
\code{impute} returns the vector \code{x} with missing values imputed via \code{method}.

\code{impute} is designed to be readable from the function call.
  For example:
\itemize{
  \item \code{impute(x, -1)} can be read as
    "\strong{impute} missing values for \strong{x} as \strong{-1}".
  \item \code{impute(x, mean)} can be read as
    "\strong{impute} missing values for \strong{x} as \strong{the mean}".
  \item \code{impute(x, sample, replace=TRUE)} can be read as
    "\strong{impute} missing values for \strong{x} using \strong{random sampling with replacement}".
}

\code{method} can either be a value, a vector of values, or a function.  Standard recycling
  is used if the vector of values is less than the number of \code{NA} values. When \code{method}
  is a function it takes only the non-\code{NA} values as input so that there is no need to
  supply those pesky \code{na.rm=TRUE} arguments. :)
}
\examples{
x <- c(NA, 1, 1, 2, 3, NA)

# Impute -1
impute(x, -1)

# Impute mean
impute(x, mean)
}

