% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/impute_ecdf.R
\name{impute_ecdf}
\alias{impute_ecdf}
\title{Helper function that samples from the empirical cumulative distribution function (ecdf)}
\usage{
impute_ecdf(x, n = 1)
}
\arguments{
\item{x}{A vector.}

\item{n}{The number of values to impute.}
}
\value{
The output of \code{impute_ecdf} is a vector of imputed values of length \code{n}
}
\description{
\code{impute} returns a vector of values imputed from sampling from the ecdf of \code{x}.
}
\examples{
x <- c(1, 2, 3, 4, 5, NA, NA)

# Impute via ecdf sampling
x_impute <- impute(x, impute_ecdf)

# Compare
data.frame(x, x_impute)
}

