% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_sample.R
\name{impute_sample}
\alias{impute_sample}
\title{Helper function that samples with replacement}
\usage{
impute_sample(x, n = 1, replace = TRUE)
}
\arguments{
\item{x}{A vector.}

\item{n}{The number of values to impute.}

\item{replace}{Whether or not to sample with replacement.  The default is TRUE.}
}
\value{
The output of \code{impute_sample} is a vector of imputed values of length \code{n}
}
\description{
\code{impute} returns a vector of values imputed from resampling \code{x}.
}
\examples{
x <- c("mode", "mode", "other", NA, NA)

# Impute via sampling
x_impute <- impute(x, impute_sample)

# Compare
data.frame(x, x_impute)
}
\seealso{
Other imputers: \code{\link{impute_ecdf}},
  \code{\link{impute_mode}}, \code{\link{impute}}
}

