% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{layer_apply_normalization}
\alias{layer_apply_normalization}
\title{Apply normalization function to input tensor}
\usage{
layer_apply_normalization(
  x,
  norm_type,
  depth = tail(shape_list2(x), 1),
  eps = 1e-09
)
}
\description{
Apply normalization function to input tensor
}
