% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attention-layers.R
\name{layer_compute_qkv}
\alias{layer_compute_qkv}
\title{Split input into query, key, value matrices in preparation for
passing to an attention layer}
\usage{
layer_compute_qkv(
  query,
  memory = NULL,
  key_depth = 64L,
  value_depth = 64L,
  q_filter_width = 1L,
  kv_filter_width = 1L,
  q_padding = "SAME",
  kv_padding = "SAME",
  vars_3d_num_heads = 0L
)
}
\description{
Uses .compute_attention_component function vie T2T framework
}
