% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attention-layers.R
\name{layer_compute_qkv_v2}
\alias{layer_compute_qkv_v2}
\title{Split input into query, key, value matrices in preparation for
passing to an attention layer}
\usage{
layer_compute_qkv_v2(x, filter_depth, num_parts = 1L, share_kv = FALSE)
}
\description{
Done a little differently than T2T framework.  TODO: (TEST THIS!)
}
