% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{layer_dense_relu_dense}
\alias{layer_dense_relu_dense}
\title{Hidden layer with RELU activation followed by linear projection.}
\usage{
layer_dense_relu_dense(
  x,
  filters,
  output_depth,
  first_kernel_size = 3L,
  second_kernel_size = 3L,
  output_activation = NULL,
  dropout = 0
)
}
\description{
Hidden layer with RELU activation followed by linear projection.
}
