% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{layer_feed_forward}
\alias{layer_feed_forward}
\title{Feed forward layer for transformer encoder}
\usage{
layer_feed_forward(
  x,
  filters,
  hidden_depth,
  dropout = 0,
  layer_type = "dense_relu_dense",
  conv_pad = "LEFT",
  pad_remover = NULL
)
}
\description{
Feed forward layer for transformer encoder
}
