% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attention-layers.R
\name{layer_normalization}
\alias{layer_normalization}
\title{R create_layer wrapper for keras LayerNormalization()}
\usage{
layer_normalization(
  object,
  axis = -1L,
  epsilon = 0.001,
  center = TRUE,
  scale = TRUE,
  beta_initializer = "zeros",
  gamma_initializer = "ones",
  beta_regularizer = NULL,
  gamma_regularizer = NULL,
  beta_constraint = NULL,
  gamma_constraint = NULL,
  trainable = TRUE,
  name = NULL
)
}
\description{
R create_layer wrapper for keras LayerNormalization()
}
