% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{layer_postprocess}
\alias{layer_postprocess}
\title{Postprocess layer output by applying a sequence of functions}
\usage{
layer_postprocess(
  layer_input,
  layer_output,
  sequence = NULL,
  dropout = 0,
  norm_type = "layer",
  name = NULL,
  v2 = FALSE
)
}
\description{
Postprocess layer output by applying a sequence of functions
}
