% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{layer_preprocess}
\alias{layer_preprocess}
\title{Preprocess layer input by applying a sequence of functions}
\usage{
layer_preprocess(
  layer_input,
  sequence = NULL,
  dropout = 0,
  norm_type = "layer",
  name = NULL,
  v2 = FALSE
)
}
\description{
Preprocess layer input by applying a sequence of functions
}
