% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attention-layers.R
\name{layer_self_attention_simple}
\alias{layer_self_attention_simple}
\title{Simplified Self attention layer
Expecting shape(x) == (batch, maxtime, units)}
\usage{
layer_self_attention_simple(
  x,
  filter_depth = 32L,
  output_depth = 64L,
  num_parts = 3L,
  dropout = 0,
  share_kv = TRUE
)
}
\description{
Simplified Self attention layer
Expecting shape(x) == (batch, maxtime, units)
}
