% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale a numeric vector}
\usage{
rescale(x, type = c("normal", "uniform"), mean = 0, sd = 1, min = 0,
  max = 1)
}
\arguments{
\item{x}{A numeric vector.}

\item{type}{A character string indicating the desired type of rescaling with \code{"normal"}
being the default.  This must be (an abbreviation of) one of the strings \code{"normal"}
or \code{"uniform"}.  See Details for more information.}

\item{mean}{The desired mean value for normal-style scaling.  Used only when \code{type="normal"}.}

\item{sd}{The desired standard deviation value for normal-style scaling.  Used only when
\code{type="normal"}.}

\item{min}{The desired minimum value for uniform-style scaling.  Used only when \code{type="uniform"}.}

\item{max}{The desired mean value for uniform-style scaling.  Used only when \code{type="uniform"}.}
}
\value{
The output of \code{rescale} is a rescaled numeric vector with the same length as \code{x}.
}
\description{
\code{rescale} returns a rescaled numeric vector.
}
\details{
\code{rescale} returns a rescaled version of the numeric vector \code{x}.  \code{NA} values in
  \code{x} are ignored during the rescaling process but preserved in the output.

\code{rescale} is designed to be readable from the function call.  For example,
  \code{rescale(x, "normal", mean=0, sd=1)} can be read as "\strong{Rescale} \strong{x} using
  a \strong{normal} transformation with \strong{mean 0} and \strong{standard deviation 1}".

The arguments \code{mean}, \code{sd}, \code{min} and \code{max} are used based on \code{type}.
  \code{rescale} offers a couple different options for \code{type}:

\itemize{
  \item \strong{normal}: The default option that rescales \code{x} using a normal transformation
    into a distribution with mean \code{mean} and standard deviation \code{sd}.  The default values
    for this rescaling are \code{mean=0} and \code{sd=1}.

The explicit formula for this transformation is:

\deqn{(x - \mu_x)/(\sigma_x) * \sigma + \mu}

where \eqn{\mu_x} is the sample mean of \code{x}, \eqn{\sigma_x} is the sample standard deviation
    of \code{x}, \eqn{\mu} is the desired mean, and \eqn{\sigma} is the desired standard deviation.

\item \strong{uniform}: rescales \code{x} using a uniform transformation into a distribution
    with minimum \code{min} and maximum \code{max}.  The default values for this rescaling are
    \code{min=0} and \code{max=1}.

The explicit formula for this transformation is:

\deqn{((x - min_x)/(max_x - min_x)) * (max - min) + min}

where \eqn{min_x} is the sample minimum of \code{x}, \eqn{max_x} is the sample maximum of \code{x},
    \eqn{min} is the desired minimum, and \eqn{max} is the desired maximum.
}
}
\examples{
set.seed(1337)
x <- rnorm(1e4)
summary(x)

# Rescale to standard normal
x_normal <- rescale(x, 'n')
summary(x_normal)
mean(x_normal); sd(x_normal)

# Rescale to standard uniform
x_uniform <- rescale(x, 'u')
summary(x_uniform)
min(x_uniform); max(x_uniform)

# Rescale to normal with mean=5 and sd=2
x_normal <- rescale(x, 'n', mean=5, sd=2)
summary(x_normal)
mean(x_normal); sd(x_normal)

# Rescale to uniform with min=-3 and max=10
x_uniform <- rescale(x, 'u', min=-3, max=10)
summary(x_uniform)
min(x_uniform); max(x_uniform)
}

