% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attention-utils.R
\name{reshape_by_blocks}
\alias{reshape_by_blocks}
\title{Reshape input by splitting length over blocks of memory_block_size.}
\usage{
reshape_by_blocks(x, x_shape, memory_block_size)
}
\description{
x Tensor [batch, heads, length, depth]
x_shape tf$TensorShape of x
memory_block_size Integer to dividing length by
Return 
  Tensor [batch, heads, length %/% memory_block_size, memory_block_size, depth]
}
