% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{sepconv_relu_sepconv}
\alias{sepconv_relu_sepconv}
\title{Hidden layer with RELU activation followed by linear projection}
\usage{
sepconv_relu_sepconv(
  x,
  filters,
  output_depth,
  first_kernel_size = list(1L, 1L),
  second_kernel_size = list(1L, 1L),
  padding = "LEFT",
  dropout = 0
)
}
\description{
Hidden layer with RELU activation followed by linear projection
}
