% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor-utils.R
\name{shape_list2}
\alias{shape_list2}
\title{Can we cheat and call value on Dimension 
class object without getting into trouble?}
\usage{
shape_list2(x)
}
\description{
Can we cheat and call value on Dimension 
class object without getting into trouble?
}
