% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attention-utils.R
\name{split_last_dimension}
\alias{split_last_dimension}
\title{Reshape x so that the last dimension becomes two dimensions.}
\usage{
split_last_dimension(x, n)
}
\description{
Reshape x so that the last dimension becomes two dimensions.
}
