% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoder.R
\name{transformer_encoder_v2}
\alias{transformer_encoder_v2}
\title{Define Transformer encoder function}
\usage{
transformer_encoder_v2(
  x,
  embedding_function = NULL,
  hparams = default_transformer_hparams()
)
}
\description{
Define Transformer encoder function
}
