% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoder.R
\name{transformer_encoder_v3}
\alias{transformer_encoder_v3}
\title{Define Transformer encoder function with lambda layer components}
\usage{
transformer_encoder_v3(
  x,
  embedding_function = NULL,
  hparams = default_transformer_hparams()
)
}
\description{
Define Transformer encoder function with lambda layer components
}
