% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformr.R
\docType{package}
\name{transformr-package}
\alias{transformr}
\alias{transformr-package}
\title{Methods for Data Transformation}
\description{
Tools for transforming data before further analysis
}
\section{Citation}{

yingjie year
}

\seealso{
Useful links:
\itemize{
  \item \url{https://yingjiehu.com}
  \item Report bugs at \url{https://github.com/huyingjie/transformr/issues}
}

}
\author{
\strong{Maintainer}: Yingjie Hu \email{yingjiehu.mail@gmail.com}

}
