% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transformr-package.R
\docType{package}
\name{transformR}
\alias{transformR}
\alias{transformR-package}
\title{The \code{transformr} package}
\description{
Functions for generating data transforms
}
\author{
Derek McCrae Norton
}

