% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/corral.R
\name{corral}
\alias{corral}
\title{Corral a character vector}
\usage{
corral(x, groups, by = c("size", "name"))
}
\arguments{
\item{x}{A character vector (or any vector than can be coerced into a character).}

\item{groups}{Either an integer with the desired number of groups or a character vector
with the groups to keep.}

\item{by}{A character string indicating the desired type of corralling with \code{"size"}
being the default.  This must be (an abbreviation of) one of the strings \code{"size"} or
\code{"name"}.  See Details for more information.}
}
\value{
The output of \code{corral} is a corralled factor vector with the same length as \code{x}.
}
\description{
\code{corral} returns a corralled factor vector.
}
\details{
\code{corral} returns a corralled version of the character vector \code{x}.  \code{NA} values in
  \code{x} are not grouped during the corralling process but preserved in the output.

\code{corral} is designed to be readable from the function call.  For example:
\itemize{
  \item \code{corral(x, groups=5, by="size")} can be read as "\strong{corral} \strong{x} into
     \strong{5} groups by \strong{size}".
  \item \code{corral(x, groups=c('a','b'), by="name")} can be read as "\strong{corral} \strong{x}
     into explicit groups for \strong{a} and \strong{b} by \strong{name}".
}
}
\examples{
set.seed(1337)
x <- sample(letters, 1e4, replace=TRUE)
summary(x)

# Corral into 5 groups by size
x_5 <- corral(x, 5, by='s')
summary(x_5)

# Corral into 5 groups by name
x_5 <- corral(x, 5, by='n')
summary(x_5)

# Corral into groups a, b, c, and others by size
x_abc <- corral(x, c('a','b','c'), by='s')
summary(x_abc)

# Corral into groups a, b, c, and others by name
x_abc <- corral(x, c('a','b','c'), by='n')
summary(x_abc)
}

