% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformr-package.R
\docType{package}
\name{transformr-package}
\alias{transformr}
\alias{transformr-package}
\title{transformr: Polygon and Path Transformations}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

In order to smoothly animate the transformation of polygons and 
    paths, many aspects needs to be taken into account, such as differing number 
    of control points, changing center of rotation, etc. The 'transformr' 
    package provides an extensive framework for manipulating the shapes of 
    polygons and paths and can be seen as the spatial brother to the 'tweenr' 
    package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasp85/transformr}
  \item Report bugs at \url{https://github.com/thomasp85/transformr/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com}

}
